python/1st/requirements.txt
--prefer-binary
Flask
emoji
tweet-preprocessor
kaggle
transformers
deep_translator
keras
numpy
pandas
sklearn
tensorflow
